# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/scheduler_postures:
    get:
      tags:
        - SchedulerPostures
      summary: List SchedulerPostures
      description: Get a paginated list of SchedulerPostures with optional filtering and sorting
      operationId: listSchedulerPostures
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerPostureList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SchedulerPostures
      summary: Create SchedulerPosture
      description: Create a new SchedulerPosture
      operationId: createSchedulerPosture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSchedulerPostureRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerPosture'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    SchedulerPosture:
      type: object
      required:
        - id
        - singleton
        - posture
        - time_rule_count
        - last_fired_count
        - metadata
      properties:
        id:
          type: string
          format: uuid
        singleton:
          type: boolean
        posture:
          type: string
        time_rule_count:
          type: integer
          format: int32
        min_delay_minutes:
          nullable: true
          type: integer
          format: int32
        interval_minutes:
          nullable: true
          type: integer
          format: int32
        last_evaluated_at:
          nullable: true
          type: string
          format: date-time
        last_fired_count:
          type: integer
          format: int32
        last_deviation_note:
          nullable: true
          type: string
          maxLength: 300
        metadata:
          type: object
...
    CreateSchedulerPostureRequest:
      type: object
      required:
        - singleton
        - posture
        - time_rule_count
        - last_fired_count
        - metadata
      properties:
        singleton:
          type: boolean
        posture:
          type: string
        time_rule_count:
          type: integer
          format: int32
        min_delay_minutes:
          nullable: true
          type: integer
          format: int32
        interval_minutes:
          nullable: true
          type: integer
          format: int32
        last_evaluated_at:
          nullable: true
          type: string
          format: date-time
        last_fired_count:
          type: integer
          format: int32
        last_deviation_note:
          nullable: true
          type: string
          maxLength: 300
        metadata:
          type: object
...
    SchedulerPostureList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SchedulerPosture'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []